+++ lib/mongo/monitor.ex
@@ -178,7 +178,7 @@ defmodule Mongo.Monitor do
         end
 
       error ->
-        Logger.warn("Unable to update server description because of #{inspect(error)}")
+        warning("Unable to update server description because of #{inspect(error)}")
         state
     end
   end
@@ -186,18 +186,8 @@ defmodule Mongo.Monitor do
   ##
   # Get a new server description from the server and send it to the Topology process.
   ##
-  defp update_server_description(%{topology_pid: topology_pid, address: address, mode: :streaming_mode} = state) do
-    case get_server_description(state) do
-      %{round_trip_time: round_trip_time} ->
-        ## debug info("Updating round_trip_time: #{inspect round_trip_time}")
-        Topology.update_rrt(topology_pid, address, round_trip_time)
-
-        %{state | round_trip_time: round_trip_time}
-
-      error ->
-        Logger.warn("Unable to round trip time because of #{inspect(error)}")
-        state
-    end
+  defp update_server_description(%{mode: :streaming_mode} = state) do
+    state
   end
 
   ##
@@ -212,7 +202,7 @@ defmodule Mongo.Monitor do
         %{state | mode: :streaming_mode, streaming_pid: pid, heartbeat_frequency_ms: 10_000}
 
       error ->
-        Logger.warn("Unable to start the streaming hello monitor, because of #{inspect(error)}")
+        warning("Unable to start the streaming hello monitor, because of #{inspect(error)}")
         state
     end
   end
@@ -283,5 +273,15 @@ defmodule Mongo.Monitor do
 
   def info(message) do
     Logger.info(IO.ANSI.format([:light_magenta, :bright, message]))
+  end
+
+  if macro_exported?(Logger, :warning, 2) do
+    defp warning(message) do
+      Logger.warning(message)
+    end
+  else
+    defp warning(message) do
+      Logger.warn(message)
+    end
   end
 end